Split out a number of embedded classes from main launchers
This should make it easier to figure out where each code section lives
instead of going back-and-forth in the same file in a jumble of mess.
